+++ docs/changelog.rst
@@ -14,6 +14,8 @@ New features
 ------------
 
 * Adding (coloured) noise to datasets (:class:`aspecd.processing.Noise`)
+* Power spectral density of 1D dataset (:class:`aspecd.analysis.PowerSpectralDensity`), *e.g.*, for analysing noise
+* Polynomial fit of 1D data (:class:`aspecd.analysis.PolynomialFit`)
 
 
 Changes